requirements.txt
python-dotenv
sqlalchemy
sqlalchemy-utils
pyodbc
pydantic